IMPALA-5582: Store sentry privileges in lower case
Privileges granted to a db/table whose name contains upper case
characters can disappear after a few seconds. A privilege
is inserted into the catalogObjectCache using a key that
uses the db/table name. The key gets converted to lower case
before inserting. The db/table name is stored in upper case in
the rolePrivilege object.
The sentryProxy thread on the other hand returns the db/table name
in lower case. If the catalogObjectCache adds the new rolePrivilege
object first and removes the old object later, it ends up deleting
the newer verison of the rolePrivilege since they both use the same
key to add/remove from the cache.
This change stores the privileges in lower cas which does not trigger
these chain on events on a sentryProxy thread update.

Change-Id: Ide3dfa601fcf77f5acc6adce9bea443aea600901