IMPALA-3897 Codegen null-aware constant in PHJ::ProcessBuildBatch()
This change codegen outs a branch in ProcessBuildBatch(). The branch never gets
executed for most of the join types except NULL_AWARE_LEFT_ANTI_JOIN.
The branch itself is not expensive to execute, but it will reduce codegen
time by removing the dead code inside the branch for almost all join
modes.

Change-Id: I06acbebc9d2d23bef4734b480a5d3ce41680ea70